backend/pyproject.toml
[project]
name = "cycle-tracker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.15.2",
    "apiflask>=2.4.0",
    "blinker>=1.9.0",
    "dotenv>=0.9.9",
    "flask>=3.1.0",
    "flask-jwt-extended>=4.7.1",
    "flask-marshmallow>=1.3.0",
    "flask-migrate>=4.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "marshmallow>=4.0.0",
    "marshmallow-sqlalchemy>=1.4.2",
    "psycopg2-binary>=2.9.10",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.0",
    "werkzeug>=3.1.3",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "factory-boy>=3.3.3",
    "isort>=6.0.1",
    "pytest>=8.3.5",
    "pytest-flask>=1.3.0",
]
